+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.2 (pre-release)
 * [Enhancement] Some fixes in French localization.
 * [Fix #33] Allow to disable the vessel overlay.
+* [Fix #35] Expose stabilization mode in API.
 
 # 2.1 (March, 3rd, 2020)
 * [Change] Switch to C# compiler v6.0.